requirements.txt
# PDF Part Extraction Pipeline Dependencies

# Web Framework
fastapi
uvicorn[standard]
python-multipart

# Core data processing
pandas
openpyxl

# OCR and document processing  
python-doctr[torch]
python-doctr[viz,html,contrib]
pymupdf

# AI and API integration
openai

# Additional utilities
Pillow
matplotlib 